qlge: update to upstream version v1.00.00.23.00.00-01
Patchwork-id: 22568
O-Subject: [RHEL6.0 PATCH] qlge: update to upstream version v1.00.00.23.00.00-01
Bugzilla: 553357
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>

This patch updates the qlge driver to a later upstream version than is
currently included in RHEL6.  This will bring the driver forward to
upstream version v1.00.00.23.00.00-01.

The patch includes all of the upstream commits through:

	commit 635267130855e91f4681df5e4d26ca9ce908447f
	Author: Ron Mercer <ron.mercer@qlogic.com>
	Date:   Sat Jan 2 10:37:44 2010 +0000

	    qlge: Add napi gro frags interface.

This patch was suggested by QLogic and they have been responsible for
testing it.

This will satisfy the request by QLogic in RHBZ 553357.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>