Expose the irq_desc node as /proc/irq/*/node
Patchwork-id: 22584
O-Subject: [RHEL6 PATCH] Expose the irq_desc node as /proc/irq/*/node
Bugzilla: 555866
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

BZ555866 Expose the irq_desc node as /proc/irq/*/node.

This patch was accepted by Andrew Morton but did not get pushed to 2.6.33
in time so SGI is asking for an exception as this is needed to complete
the FutureFeature BZ464201 IRQ Forwarding.  It is currently in
http://userweb.kernel.org/~akpm/mmotm/ as
kernel-irq-procc-expose-the-irq_desc-node-in-proc-irq.patch

This file provides device hardware locality information for apps desiring
to include hardware locality in irq mapping decisions.

NOTE: this patch does not alter the functioning of the kernel except
to provide information to user level processes so it is very safe.

The patch applies cleanly and builds against 2.6.32-6.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>