dccp: modify how dccp creates slab caches to prevent bug halt in SLUB
Patchwork-id: 22634
O-Subject: [RHEL 6 PATCH] dccp: modify how dccp creates slab caches to prevent
	bug halt in SLUB (bz 553698)
Bugzilla: 553698
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

Hey all-
	We currently in RHEL6 get a bug halt when we unload the dccp module.
This stems from the fact that dccp kmallocs data to hold the name of the caches
it creates.  The slub implementation of kmem_cache_create has a 'merging'
feature, in which caches of simmilar configurations wind up using the same data
structure.  When/if such a merge occurs, one of the two cache names (the last
one to be created) is dropped by slub. As such, when kmem_cache_name is called
to retrieve the name of a cache, it may be different than the name that was
passed in during kmem_cache_create.  Since dccp uses this call to retrieve the
name of the cache on destroy, it gets an unexpected pointer (usually something
pointing into the static string table), which it then passes to kfree, which in
turn trips on the BUG halt, after failing the !PageSlab test in the slub kfree
function.

	Normally this isn't a problem, since creators of kmem caches never alloc
or free their names, using (static strings instead).  This seems like a bug in
slub, but fixing it there invites some complexity.  This patch instead just
converts the storage of cache names in dccp to be a static part of the
ccid_operations structure.  It doesn't cost any extra memory, and saves us
having to do a kstrudup per cache during init.  I've posted it upstream and its
gotten an ack from acme so far (who maintains dccp).

Fixes bz 553698.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>