Disable Memory hot add on x86 32-bit
Patchwork-id: 22677
O-Subject: [RHEL6]: x86: Disable Memory hot add on x86 32-bit
Bugzilla: 557131
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Christopher Lalancette <clalance@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Peter Martuccelli <peterm@redhat.com>

Memory hot add currently does not work on 32-bit.  The kernel panics or hangs
when attempting to add memory.

Patches are currently being worked on upstream and are unlikely to be ready
for the RHEL6.0 deadline.

We cannot disable MEMORY_HOTPLUG, and re-enable it later because of kabi
concerns.

Disable memory hot add for x86 32-bit RHEL6.0.  We will re-enable it for
RHEL6.1.

Resolves BZ 557131.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>