e1000e: update to the latest upstream
Patchwork-id: 22682
O-Subject: [RHEL6.0 PATCH] e1000e: update to the latest upstream
Bugzilla: 462780
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

This is an e1000e update to the latest that is in Linus' tree, so it
should be what ships with 2.6.33.  The last upstream commit included in
this set is:

	commit baf86c9d36826fab0160251bbc87dfab3af48a21
	Author: Bruce Allan <bruce.w.allan@intel.com>
	Date:   Wed Jan 13 01:53:08 2010 +0000

	    e1000e: workaround link issues on busy hub in half duplex on 82577/82578

Included are several important bug fixes (including the one posted
earlier this week and last week for RHEL5) and code to enable the
82567V-3 devices is included.

This driver is almost identical to upstream (the only difference being
the lack of the netdev_alloc_skb_ip_align function and the changes
needed to accommodate the addition).

I did some regression testing on this driver and found no problems.

This will resolve the request in RHBZ 462780.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>