kABI internal only files
Patchwork-id: 22779
O-Subject: Re: [RHEL6] kABI (internal only) files
Bugzilla:

My test build succeeded on all platforms (with the exception that the
SPEC file is currently disabling it on s390x - that is a simple toggle -
and I will just for the sake of it, re-enable and rebuild on there later
on) and I have attached the patch that I made to the current git tree.

Some points about the current kABI tests:

1). There's a deprecation warning about the sha library. I will fix that
by converting it over (the kabitool script) and send you a patch.

2). I will perhaps send a patch to silence the warning about a lack of
Module.kabi file on those builds where there is no kABI checking (the
debug builds and kdump), though that's only a harmless message that
existed for a long time in RHEL5 anyway.

Jon.
commit f69d188778ac11c3a72f50c851430ce47715a0db
Author: Jon Masters <jcm@redhat.com>
Date:   Sun Jan 24 00:05:04 2010 -0500

    kabi: Update kABI metadata for beta1

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>