io-apic: Move the effort of clearing remoteIRR explicitly before migrating the irq
Patchwork-id: 22786
O-Subject: [PATCH RHEL6 6/9 BZ 463633] x86,
	io-apic: Move the effort of clearing remoteIRR explicitly before
	migrating the irq
Bugzilla: 463633
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

x86, io-apic: Move the effort of clearing remoteIRR explicitly before migrating the irq

ca64c47cecd0321b2e0dcbd7aaff44b68ce20654
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=ca64c47cecd0321b2e0dcbd7aaff44b68ce20654

When the level-triggered interrupt is seen as an edge interrupt,
we try to clear the remoteIRR explicitly (using either an
io-apic eoi register when present or through the idea of
changing trigger mode of the io-apic RTE to edge and then back
to level). But this explicit try also needs to happen before we
try to migrate the irq. Otherwise irq migration attempt will
fail anyhow, as it postpones the irq migration to a later
attempt when it sees the remoteIRR in the io-apic RTE still set.

Signed-off-by: "Maciej W. Rozycki" <macro@linux-mips.org>
Reviewed-by: Suresh Siddha <suresh.b.siddha@intel.com>
Cc: ebiederm@xmission.com
Cc: garyhade@us.ibm.com
LKML-Reference: <20091201233334.975416130@sbs-t61.sc.intel.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>