config: disable CONFIG_X86_CPU_DEBUG
Patchwork-id: 22834
O-Subject: [RHEL6 PATCH] config: disable CONFIG_X86_CPU_DEBUG
Bugzilla: 558597
RH-Acked-by: Danny Feng <dfeng@redhat.com>
RH-Acked-by: Dave Anderson <anderson@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=558597

This patch disables CONFIG_X86_CPU_DEBUG a feature that will be removed soon
from upstream based on upstream discussion:
	http://lkml.org/lkml/2010/1/13/351
as pointed by Dave Jones. There's no specific request for this option to be
enabled and it's currently compiled as module because it was enabled in Fedora.

This patch was tested with 'make rh-configs-prep'.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>