Memory tracking for Stratus
Patchwork-id: 22971
O-Subject: [RHEL6 PATCH] Memory tracking for Stratus w/fewer exports
Bugzilla: 512400
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=512400

In response to Christoph's comment, I took a closer look and it turns out
that some of the symbol exports are not necessary after all; they're used
in the kernel but not in the client code.  Attached is a revised patch.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>