zcrypt: Do not remove coprocessor in case of error 8/72
Patchwork-id: 23000
O-Subject: [RHEL6 PATCH 2/3] [s390] zcrypt: Do not remove coprocessor in case of
	error 8/72
Bugzilla: 559613
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------

In a case where the number of the input data is bigger than the
modulus of the key, the coprocessor adapters will report an
rc=8 reason=72 error.  This case is not caught yet, thus the
adapter will be taken offline.

To prevent this, return an -EINVAL instead.

Bugzilla
--------
BZ 559613
https://bugzilla.redhat.com/show_bug.cgi?id=559613

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=19b123ebacacdce5e75045bfe82122b01c821a5b

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>