GFS2: Wait for unlock completion on umount
Patchwork-id: 23110
O-Subject: [RHEL6] GFS2: Wait for glocks on umount (bz #561287) [1/2]
Bugzilla: 561287
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Benjamin Marzinski <bmarzins@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Hi,

This patch and the following patch together ensure that we always wait for
all replies to demote requests from the DLM at umount time. This prevents
the gfs2 module from being removed until all the replies have been
received.

The patches have been tested by Fabio and are also in the upstream GFS2
-nmw and -fixes git trees. I'm intending to send them to Linus shortly,

Steve.

>From e402746a945ceb9d0486a8e3d5917c9228fa4404 Mon Sep 17 00:00:00 2001
From: Steven Whitehouse <swhiteho@redhat.com>
Date: Mon, 25 Jan 2010 11:20:19 +0000
Subject: [PATCH 1/2] GFS2: Wait for unlock completion on umount

This patch adds a wait on umount between the point at which we
dispose of all glocks and the point at which we unmount the
lock protocol. This ensures that we've received all the replies
to our unlock requests before we stop the locking.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Reported-by: Fabio M. Di Nitto <fdinitto@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>