sunrpc: parse and return errors reported by gssd
Patchwork-id: 23193
O-Subject: [RHEL6 PATCH 1/4] BZ#479359: sunrpc: parse and return errors reported
	by gssd
Bugzilla: 479359
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

The kernel currently ignores any error code sent by gssd and always
considers it to be -EACCES. In order to better handle the situation of
an expired KRB5 TGT, the kernel needs to be able to parse and deal with
the errors that gssd sends. Aside from -EACCES the only error we care
about is -EKEYEXPIRED, which we're using to indicate that the upper
layers should retry the call a little later.

To maintain backward compatibility with older gssd's, any error other
than -EKEYEXPIRED is interpreted as -EACCES.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>