nfs: handle NFSv3 -EKEYEXPIRED errors as we would -EJUKEBOX
Patchwork-id: 23195
O-Subject: [RHEL6 PATCH 3/4] BZ#479359: nfs: handle NFSv3 -EKEYEXPIRED errors as
	we would -EJUKEBOX
Bugzilla: 479359
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

We're using -EKEYEXPIRED to indicate that a krb5 credcache contains an
expired ticket and that we should have the NFS layer retry the RPC call
instead of returning an error back to the caller. Handle this as we
would an -EJUKEBOX error return.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>