virtio_blk: add block topology support
Patchwork-id: 23235
O-Subject: [RHEL6 PATCH] virtio_blk: add block topology support
Bugzilla: 556477
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

This is the kernel side of BZ 556477 - backport the topology support for
the virtio-blk driver which was accepted for 2.6.34 by Rusty.

Note that I couldn't test it in RHEL6 as the redhat-private or patches
build system is too messed up and required random bytes to build which
my remote testbox simply doesn't have.
Allow reading various alignment values from the config page.  This
allows the guest to much better align I/O requests depending on the
storage topology.

Note that the formats for the config values appear a bit messed up,
but we follow the formats used by ATA and SCSI so they are expected in
the storage world.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>