zfcp: cancel all pending work for a to be removed zfcp_port
Patchwork-id: 23330
O-Subject: [RHEL6 PATCH 1/5] [s390] zfcp: cancel all pending work for a to be
	removed zfcp_port
Bugzilla: 564382

Description
-----------
In some cases an invalid port cannot be detached because a pending
work is already queued which is holding a reference to this port.
This results in stalled queues and processes, finally, leading to an
unusable adapter.

To avoid this situation ensure no work is pending on the same queue
on which a port_dequeue is currently running.

Bugzilla
--------
BZ 564382
https://bugzilla.redhat.com/show_bug.cgi?id=564382

Upstream status of the patch
----------------------------
The patch is specific for the 2.6.32 kernel.  The problem is solved upstream
(2.6.33) by changes in locking and reference counting.  Backporting these
changes would require additional development and test efforts.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>