zfcp: report BSG errors in correct field
Patchwork-id: 23329
O-Subject: [RHEL6 PATCH 2/5] [s390] zfcp: report BSG errors in correct field
Bugzilla: 564378

Description
-----------
Errors are reported even though the request is processed successfully
because the wrong variables are used to report the execution status.

The status FC_CTELS_STATUS_REJECT for all FC BSG errors is not appropriate.
Instead, report -EIO in the result field if there was a problem in zfcp
with the FC BSG request.  If the request is good from our point of view,
report result 0, status FC_CTELS_STATUS_OK and let userspace read the
Accept or Reject from the payload (as documented in scsi_bsg_fc.h).

Bugzilla
--------
BZ 564378
https://bugzilla.redhat.com/show_bug.cgi?id=564378

Upstream status of the patch
----------------------------
The patch will be upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-rc-fixes-2.6.git;a=commitdiff;h=7dec9cf1dfa283feca4b761160112ea4838a6a8c

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.